+++ src/contexts/Farming/Context.ts
@@ -5,6 +5,7 @@ import { ContextValues } from "./types";
 const Context = createContext<ContextValues>({
   farmingStartTimes: [1],
   setConfirmModal: () => {},
+  inkTotalSupply: 0,
 });
 
 export default Context;